An experimental repo for an AI Coding Chatbot.
NOTE: This is meant to be platform agnostic, but for experimentation purposes I've used Ruulf's freeGPT Python Package.

`coding_copilot.py`, when run, initiates a chatbot that connects to GPT-3.5 and immediately extracts any code written by the AI into a real Python file so the user can test it immediately.